+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [3.8.1] - 2024-05-10
+
 ### Fixed
 
 - `AuditClient.logStream()` Javadoc parameter name.